+++ config/reverb.php
@@ -43,10 +43,10 @@ return [
                 'server' => [
                     'url' => env('REDIS_URL'),
                     'host' => env('REDIS_HOST', '127.0.0.1'),
-                    'port' => env('REDIS_PORT', 6379),
+                    'port' => (int) env('REDIS_PORT', 6379),
                     'username' => env('REDIS_USERNAME'),
                     'password' => env('REDIS_PASSWORD'),
-                    'database' => env('REDIS_DB', 0),
+                    'database' => (int) env('REDIS_DB', 0),
                     'timeout' => env('REDIS_TIMEOUT', 60),
                 ],
             ],
